updated name and some conditions
